Added some metadata to disable the cache.
This code was given to me by Sean Wheeler, to disable the cache feature. I found that I'd launch a server but no matter what changes I made to the CSS file, it had no effect on the page. These 3 lines of metadata are aimed to fix this problem. They will be disabled before I submit this project.